refactor: BaseEntity 제거 후 공통 필드 각 엔티티에 직접 선언, 각 엔티티에 Serializable 구현
