--- 2-sprint-mission/src/main/java/com/sprint/mission/discodeit/entity/BaseEntity.java
@@ -1,31 +1,0 @@
-package com.sprint.mission.discodeit.entity;
-
-import java.util.UUID;
-
-public abstract class BaseEntity {
-    private final UUID id;
-    private final long createdAt;
-    private long updatedAt;
-
-    protected BaseEntity() {
-        this.id = UUID.randomUUID();
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = this.createdAt;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public long getCreatedAt() {
-        return createdAt;
-    }
-
-    public long getUpdatedAt() {
-        return updatedAt;
-    }
-
-    protected final void updateTimestamp() {
-        this.updatedAt = System.currentTimeMillis();
-    }
-}
