+++ 2-sprint-mission/src/main/java/com/sprint/mission/discodeit/entity/Message.java
@@ -1,17 +1,36 @@
 package com.sprint.mission.discodeit.entity;
 
+import java.io.Serializable;
 import java.util.UUID;
 
-public class Message extends BaseEntity {
+public class Message implements Serializable {
+    private final UUID id;
+    private final long createdAt;
+    private long updatedAt;
     private final UUID userId;
     private final UUID channelId;
     private String content;
+    private static final long serialVersionUID = 1L;
 
     public Message(UUID userId, UUID channelId, String content) {
-        super();
+        this.id = UUID.randomUUID();
+        this.createdAt = System.currentTimeMillis();
+        this.updatedAt = this.createdAt;
         this.userId = userId;
         this.channelId = channelId;
         this.content = content;
+    }
+
+    public UUID getId() {
+        return id;
+    }
+
+    public long getCreatedAt() {
+        return createdAt;
+    }
+
+    public long getUpdatedAt() {
+        return updatedAt;
     }
 
     public UUID getUserId() {
@@ -31,10 +50,14 @@ public class Message extends BaseEntity {
         updateTimestamp();
     }
 
+    protected final void updateTimestamp() {
+        this.updatedAt = System.currentTimeMillis();
+    }
+
     @Override
     public String toString() {
         return "Message{" +
-                "id=" + getId() +
+                "id=" + id +
                 ", content='" + content + '\'' +
                 ", userId=" + userId +
                 ", channelId=" + channelId +
